Scaling diagram fix, encapsulation work
Reworked the drawing of the graph to look proper when using non-integer
displayScale values (higher than 1).

The diagram and probability drawing are no longer using the constants
but rely on input parameters, a first step to encapsulating things for
reusability (like drawing two graphs in the same diagram).

An override for the highest y axis value has also been added.