Breaking a few things for nice pictures
I guess that says it all. I used a white background and disabled the
diagram drawing to isolate the line plots for the other two versions
(see http://playdota.com/forums/showpost.php?p=9626435&postcount=138 )
and then broke some more things to make the diagram look decent.

Layering of the images (included in the commit) will be done manually.
It's not worth the effort to code all this in here, especially since the
parts/values/behaviors that should be adjustable without digging in code
are unclear.